move function to utils/message.js Optimize the code
