+++ src/components/Widget/components/Conversation/components/Messages/components/Message/styles.scss
@@ -4,6 +4,8 @@
 .rcw-message {
   margin: 10px;
   display: flex;
+  font-size: 14px;
+  word-break: break-all;
 }
 
 .rcw-client {
@@ -13,10 +15,6 @@
 
 .rcw-response {
   @include message-bubble($grey-2);
-}
-
-.rcw-message {
-  font-size: 14px;
 }
 
 /* For markdown elements created with default styles */
