+++ src/components/Widget/components/Conversation/components/Messages/styles.scss
@@ -7,6 +7,7 @@
   max-height: 410px;
   overflow-y: scroll;
   padding-top: 10px;
+  -webkit-overflow-scrolling: touch;
 }
 
 .rcw-full-screen {
