+++ src/utils/messages.js
@@ -1,5 +1,5 @@
 import { Map } from 'immutable';
-import { MESSAGES_TYPES, MESSAGE_SENDER } from '@constants';
+import { MESSAGES_TYPES, MESSAGE_SENDER, MESSAGE_BOX_SCROLL_DURATION } from '@constants';
 
 import Message from '@messagesComponents/Message';
 import Snippet from '@messagesComponents/Snippet';
@@ -34,4 +34,48 @@ export function createComponentMessage(component, props, showAvatar) {
     sender: MESSAGE_SENDER.RESPONSE,
     showAvatar
   });
+}
+
+/**
+ * Easing Functions
+ * @param {*} t timestamp
+ * @param {*} b begining
+ * @param {*} c change
+ * @param {*} d duration
+ */
+function sinEaseOut(t, b, c, d) {
+  return c * ((t = t / d - 1) * t * t + 1) + b;
+}
+
+/**
+ * 
+ * @param {*} target scroll target
+ * @param {*} scrollStart
+ * @param {*} scroll scroll distance
+ */
+function scrollWithSlowMotion(target, scrollStart, scroll) {
+  const raf = window.webkitRequestAnimationFrame || window.requestAnimationFrame
+  let start = null
+  const step = (timestamp) => {
+    if (!start) {
+      start = timestamp
+    }
+    let stepScroll = sinEaseOut(timestamp - start, 0, scroll, MESSAGE_BOX_SCROLL_DURATION)
+    let total = scrollStart + stepScroll
+    target.scrollTop = total;
+    if (total < scrollStart + scroll) {
+      raf(step)
+    }
+  }
+  raf(step)
+}
+
+export function scrollToBottom(messagesDiv) {
+  if (!messagesDiv) return;
+  const screenHeight = messagesDiv.clientHeight;
+  const scrollTop = messagesDiv.scrollTop;
+
+  const scrollOffset = messagesDiv.scrollHeight - (scrollTop + screenHeight)
+
+  scrollOffset && scrollWithSlowMotion(messagesDiv, scrollTop, scrollOffset);
 }
